+++ sheet_service.py
@@ -60,7 +60,7 @@ def update_google_sheet(service, spreadsheet_id, range_to_update, rows):
   
   return response
 
-def batch_update_google_sheet(service, spreadsheet_id, rows):
+def batch_update_objective_rubric(service, spreadsheet_id, rows):
     major_dim = "COLUMNS"
     start_row_num = 12
     end_row_num = start_row_num + len(rows[0])
@@ -109,3 +109,48 @@ def batch_update_google_sheet(service, spreadsheet_id, rows):
     request = service.spreadsheets().values().batchUpdate(spreadsheetId=spreadsheet_id, body=body)
     response = request.execute()
     return response
+
+
+def create_new_sheet(service, name):
+  spreadsheet_body = {
+    "properties": {
+      "title": name
+    }
+  }
+
+  request = service.spreadsheets().create(body=spreadsheet_body)
+  response = request.execute()
+  return response
+
+def push_csv_to_gsheet(service, csv_path, spreadsheet_id, sheet_id):
+  """Pushes a csv file to google sheets api.
+
+  Args:
+      service ([Google resource]): [resource for accessing sheets api]
+      csv_path ([string]): [filepath for csv file to be uploaded]
+      spreadsheet_id ([string]): [id of spreadsheet to upload file]
+      sheet_id ([string]): [id for which worksheet to upload data]
+
+  Returns:
+      [Google Spreadsheet]: [Spreadsheet includes id, sheets]
+  """  
+  API = service
+  with open(csv_path, 'r') as csv_file:
+      csvContents = csv_file.read()
+  body = {
+      'requests': [{
+          'pasteData': {
+              "coordinate": {
+                  "sheetId": sheet_id,
+                  "rowIndex": "0",  # adapt this if you need different positioning
+                  "columnIndex": "0", # adapt this if you need different positioning
+              },
+              "data": csvContents,
+              "type": 'PASTE_NORMAL',
+              "delimiter": ',',
+          }
+      }]
+  }
+  request = API.spreadsheets().batchUpdate(spreadsheetId=spreadsheet_id, body=body)
+  response = request.execute()
+  return response
